pyproject.toml
[tool.poetry]
name = "ai-db-agent"
version = "0.1.0"
description = ""
authors = ["Connor (Cub) <connor@cubdigital.co.nz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
langchain = "^0.2.14"
python-dotenv = "^1.0.1"
gradio = "^4.42.0"
pyodbc = "^5.1.0"
langchain-openai = "^0.1.22"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
